+++ Source/Samples/SamplesManager.cpp
@@ -117,7 +117,6 @@
 #include "53_LANDiscovery/LANDiscovery.h"
 #endif
 #include "54_WindowSettingsDemo/WindowSettingsDemo.h"
-#include "99_RayCast/RayCastSample.h"
 #if URHO3D_SYSTEMUI
 #include "100_HelloSystemUI/HelloSystemUI.h"
 #endif
@@ -143,6 +142,7 @@
 #if URHO3D_RMLUI
 #include "114_AdvancedUI/AdvancedUI.h"
 #endif
+#include "115_RayCast/RayCastSample.h"
 #include "Rotator.h"
 
 #include "SamplesManager.h"
@@ -360,7 +360,6 @@ void SamplesManager::Start()
     RegisterSample<LANDiscovery>();
 #endif
     RegisterSample<WindowSettingsDemo>();
-    RegisterSample<RayCastSample>();
 #if URHO3D_SYSTEMUI
     RegisterSample<HelloSystemUi>();
 #endif
@@ -386,6 +385,7 @@ void SamplesManager::Start()
 #if URHO3D_RMLUI
     RegisterSample<AdvancedUI>();
 #endif
+    RegisterSample<RayCastSample>();
 
     if (!commandLineArgs_.empty())
         StartSample(commandLineArgs_[0]);
